Add support for CA.
Self signed certificates are very easy to setup but they have two main
problems.  The first is that certificate handling does not scale when multiple
client TLS certificates must be created for a single server.  The second is
that Chromium seems to refuse self signed certificates, even when the server
certificate is added to the list of trusted certificates (see
https://bugs.pcbsd.org/issues/6864).

The solution is to add support for private Certificate Authority (CA).

Add support for generating certificates as a Certificate Authority, and to sign
server and client certificates using the CA.

Improve documentation.